v2.0.2 novaj_bibliotekoj2 Freŝa versio de DR Radio kaj multaj el la bibliotekoj

+++ EoRadio/app/src/main/java/dk/dr/radio/data/HentedeUdsendelser.java
@@ -281,7 +281,7 @@ public class HentedeUdsendelser {
     tjekDataOprettet();
     HentetStatus hs = getHentetStatus(u);
     if (hs.startUri!=null) new File(URI.create(hs.startUri)).delete(); // Hvis ikke hentet færdig endnu er hs.startUri==null
-    new File(hs.destinationFil).delete();
+    if (hs.destinationFil!=null) new File(hs.destinationFil).delete();
 
     data.hentetStatusFraSlug.remove(u.slug);
     Long id = data.downloadIdFraSlug.remove(u.slug);
